requirements.txt
numpy==1.23.5
pillow
opencv-python
imagehash
tqdm
pandas
scikit-learn==1.2.2
albumentations==1.3.1
tensorflow==2.12.0
tensorflow-addons==0.20.0
matplotlib
seaborn
flask
tensorflow
keras
torch
numpy
scikit-learn
